Moved fileStorage.js
